restructuring cards related scripts to cards.py
